requirements.txt
moderngl
moderngl-window
PIL
pyglm
cson
overrides
pycairo
pytweening
assign
ptpython
watchdog
pyogg
pyopenal
webcolors
